Split generation orchestration tests
Move shared orchestration fakes and builders into a non-test helper module.
Split DTO, planner, orchestrator, show-notes executor, and property tests into
focused modules so the original broad test file can be removed.
